Move badge to badges folder
